created common file name function for USFM export and alignment export.
